backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
python-multipart>=0.0.12
psycopg2-binary>=2.9.10
email-validator>=2.2.0
pypdf>=5.1.0
python-docx>=1.1.2
